Adjusted Navigation Bar CSS; Cleaned up achievements.html
I set the navigation bar with the Achievements and Quests buttons at the top for every page. For some reason, the nav bar was hovering over the calendar in index.html, so I fixed that. I also created achievements.css to classify repeated styling for the progress bars and claim buttons and adjusted achievements.html accordingly.